+++ .github/workflows/npm-publish.yml
@@ -19,6 +19,7 @@ jobs:
           node-version: 18
           registry-url: https://registry.npmjs.org/
       - run: npm i
+      - run: npm run build
       - run: npx semantic-release
         env:
           NODE_AUTH_TOKEN: ${{secrets.npm_token}}
