+++ .gitignore
@@ -62,3 +62,4 @@ typings/
 
 #webstorm
 .idea
+dist
